Fix unequal sized Dock items
When content padding is applied before the onMeasure phase of the view a
padding is set to the view. This happened for dynamic apps that set non
zero content padding at startup. To fix, the content padding is posted
on the queue which would garuntee that it is called after onMeasure
phase.

Bug: 320542080
Bug: 301482374
Test: manual, atest CarDockLibTests

Change-Id: Ibb30e672bbba60603af2a1afa738b90c62485a50