Move InstantTaskExecutorRule to test package and eliminate uses of the version from car-apps-common.
`car-apps-common` improperly included the rule in it's main (non-test) code.  It's being refactored out to `car-testing-common`. Launcher already had a copy of the file, but in the wrong directory, so that file is being moved to the app/tests directory and the tests are having their includes changed to use that version.

Bug: 389088738
Test: mma CarLauncherTests
Flag: EXEMPT refactor
Change-Id: I597a546dcbe3ab872028648ff961b66b24ade73b