Add support for App Widgets
This commit introduces the functionality to host App Widgets within the Car Launcher application.

A new `WidgetHostActivity` is added, which serves as a container for displaying a list of App Widgets. This activity utilizes `AppWidgetHost` to manage the lifecycle of the widgets.

To demonstrate this functionality, a sample "Date" widget (`DateAppWidgetProvider`) is included. This widget displays the current date and day of the week.

The entire App Widget host functionality is controlled by the feature flag `app_widget_host`.

Bug: 457869713
Test: manual
Flag: com.android.car.carlauncher.app_widget_host
Change-Id: I49a799f3251d8861a4acc243cb05b4b778fa4410